[Platinum V] Title: 볼록 껍질, Time: 44 ms, Memory: 5148 KB -BaekjoonHub

+++ 백준/Platinum/1708. 볼록 껍질/볼록 껍질.cc
@@ -3,7 +3,7 @@ using namespace std;
 
 struct POINT{
     long long x, y;
-    int x_d, y_d;
+    long long x_d, y_d;
 }arr[100001];
 
 bool com(POINT &p, POINT &q){
